Working rewind and resume
Fixed an issue with the CharacterController where after a restore it would still think its in the old position. Disable and enable it around the restore to fix this. Fixed animation interpolation by removing the complicated looping logic.

The last step with the rewind demo, other than some cleanup, is to get multiple rewinds to work. Currently if you pause, rewind, resume, then pause again you see a lot of errors.  One thing I think needs doing is that when we resume it resets the WRITE head to the point at which we resumed, and also modifies the recorded frame count.