Data schemas and fixes for replay
Added the concept of data schemas for the rewind components.
Each component now has to define its schema, which is basically just a list of the data it writes out.
Right now this is only used to replace the data size calculation for each component, but could be used for validation and to help with serialization.
Some fixes for vfx material chagnes in replays.